Return integer reception quantities

+++ backend/IQBF.Application/DTOs/Receptions/ReceptionDto.cs
@@ -1,10 +1,6 @@
 namespace IQBF.Application.DTOs.Receptions;
 
-public sealed record ReceptionItemDto(
-    Guid BLId,
-    string BLCode,
-    decimal Quantity
-);
+public sealed record ReceptionItemDto(Guid BLId, string BLCode, int Quantity);
 
 public sealed record ReceptionDto(
     Guid Id,
